pyproject.toml
[build-system]
requires = [
	"flit_core >=3.2,<4",
	"python_version >= '3.7'"
	]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[project]
name = "search"
authors = [{name = "Han Chen", email = "han.chen@ucsf.edu"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "pytest",
	"scipy.spatial.distance",
	"random"
]

[project.urls]
Home = "https://github.com/hchen725/HW3-Clustering"